requirements-prod.txt
# Production requirements for AI Wiki Quiz Generator
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0

# AI/ML
langchain==0.0.350
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Utilities
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Security
cryptography==41.0.7

# Monitoring
prometheus-client==0.19.0

# CORS
python-cors==1.7.0
